refactor: change author_id and genre_id to BigSerial for Long support

+++ sql/09-functions.sql
@@ -1,12 +1,12 @@
 -- Description:
 -- Returns all books by a specific author.
 -- Parameter:
--- 		p_author_id (INT) -> ID of the author whose books we want to retrieve.
+-- 		p_author_id (BIGINT) -> ID of the author whose books we want to retrieve.
 -- Returns:
 -- 		A table with columns:
 -- 		isbn -> ISBN of the book
 -- 		title -> Title of the book
-CREATE OR REPLACE FUNCTION get_books_by_author(p_author_id INT)
+CREATE OR REPLACE FUNCTION get_books_by_author(p_author_id BIGINT)
 RETURNS TABLE(isbn TEXT, title TEXT) AS $$
 BEGIN
     RETURN QUERY
